requirements.txt
asgiref==3.2.7
boto3==1.13.25
botocore==1.16.25
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
oauthlib==3.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
